+++ lib/store/src/autoTitle.ts
@@ -48,27 +48,37 @@ function pathJoin(paths: string[]): string {
   return paths.join('/').replace(slashes, '/');
 }
 
-export const autoTitleFromSpecifier = (fileName: string, entry: NormalizedStoriesSpecifier) => {
+export const userOrAutoTitleFromSpecifier = (fileName: string, entry: NormalizedStoriesSpecifier, userTitle?: string) => {
   const { directory, importPathMatcher, titlePrefix = '' } = entry || {};
   // On Windows, backslashes are used in paths, which can cause problems here
   // slash makes sure we always handle paths with unix-style forward slash
   const normalizedFileName = slash(fileName);
 
   if (importPathMatcher.exec(normalizedFileName)) {
-    const suffix = normalizedFileName.replace(directory, '');
-    const titleAndSuffix = slash(pathJoin([titlePrefix, suffix]));
-    let path = titleAndSuffix.split('/');
-    path = stripExtension(path);
-    path = removeRedundantFilename(path);
-    return path.join('/');
+    if (!userTitle) {
+      const suffix = normalizedFileName.replace(directory, '');
+      const titleAndSuffix = slash(pathJoin([titlePrefix, suffix]));
+      let path = titleAndSuffix.split('/');
+      path = stripExtension(path);
+      path = removeRedundantFilename(path);
+      return path.join('/');
+    }
+
+    if (!titlePrefix) {
+      return userTitle;
+    }
+
+    return slash(pathJoin([titlePrefix, userTitle]));
   }
+
   return undefined;
 };
 
-export const autoTitle = (fileName: string, storiesEntries: NormalizedStoriesSpecifier[]) => {
+export const userOrAutoTitle = (fileName: string, storiesEntries: NormalizedStoriesSpecifier[], userTitle?: string) => {
   for (let i = 0; i < storiesEntries.length; i += 1) {
-    const title = autoTitleFromSpecifier(fileName, storiesEntries[i]);
+    const title = userOrAutoTitleFromSpecifier(fileName, storiesEntries[i], userTitle);
     if (title) return title;
   }
-  return undefined;
+
+  return userTitle || undefined;
 };