+++ lib/ui/src/components/preview/preview.mockdata.tsx
@@ -46,8 +46,8 @@ export const previewProps: PreviewProps = {
   queryParams: {},
   options: {
     isFullscreen: false,
-    isToolshown: true,
     showTabs: true,
+    showToolbar: true,
   },
   withLoader: false,
   docsOnly: false,